Clear ruff debt and make the lint gate repo-wide
Phase 1 of the docloom -> docsynth rename: the rename touches ~120 of 134 .py
files, and the changed-files ruff gate would then surface ~279 pre-existing
findings and block the merge. Clear the debt first so the rename PR is trivially
green.

- `ruff check --fix` for the auto-fixable set (unused-noqa, import sorting,
  unused imports, dunder-all sorting, quoted annotations, …).
- Manual fixes for the rest: merged isinstance calls, `key in dict`-style
  (noqa'd where the row is a sqlite3.Row, not a dict), redundant int() casts,
  a lowercase local, a duplicate test stub renamed, unused unpacks, a ClassVar
  annotation, semicolon statements split, and long lines wrapped.
- Scope-ignore rules that are churn, not bugs: N817/N818 (import/exception
  naming), RUF001/002/003 (intentional – / → glyphs); and E501 in the bilingual
  catalogue data, inline-SVG stamps, dev scripts, and tests.
- Flip the CI lint job from changed-files to a repo-wide `ruff check .` (pinned
  ruff) now that the tree is clean.

Full suite green (the 4 "needs gcp extra" tests fail only locally, where the
gcp extra is installed; they pass on a clean runner).

+++ src/docloom/core/pipeline/pdf.py
@@ -35,8 +35,8 @@ import math
 from typing import Any
 
 from docloom.core.enums import DocumentCondition
-from docloom.core.pack import DocumentPack
 from docloom.core.logging import get_logger
+from docloom.core.pack import DocumentPack
 from docloom.core.pipeline.degrade import degrade_pdf
 from docloom.core.pipeline.renderer import RenderedDocument
 from docloom.core.record import GoldenRecord
@@ -217,7 +217,7 @@ class PdfRenderer:
         """
         page = self._ensure_browser().new_page()
         w, h = self._printable_size()
-        page.set_viewport_size({"width": max(int(round(w)), 1), "height": max(int(round(h)), 1)})
+        page.set_viewport_size({"width": max(round(w), 1), "height": max(round(h), 1)})
         # print media so the archetypes' @print rules (table-header-group,
         # break-inside) take effect, and backgrounds so accent/zebra show.
         page.emulate_media(media="print")
